requirements.txt
textdiscord.py==2.3.2
google-generativeai==0.8.2